test: share integration helpers
