+++ tests/integration.rs
@@ -2,47 +2,15 @@ use assert_cmd::Command;
 use filetime::FileTime;
 use lsplus::utils::icons::Icon;
 use std::fs;
-use std::path::Path;
 use std::time::{Duration, SystemTime};
 use strip_ansi_escapes::strip_str;
 use tempfile::tempdir;
 
-fn run_and_capture(cmd: &mut Command) -> (String, String) {
-    let output = cmd.output().unwrap();
-    assert!(
-        output.status.success(),
-        "command failed with stderr: {}",
-        String::from_utf8_lossy(&output.stderr)
-    );
+mod common;
 
-    let stdout = strip_str(String::from_utf8_lossy(&output.stdout)).to_owned();
-    let stderr = strip_str(String::from_utf8_lossy(&output.stderr)).to_owned();
-    (stdout, stderr)
-}
-
-fn run_and_capture_raw(cmd: &mut Command) -> (String, String) {
-    let output = cmd.output().unwrap();
-    assert!(
-        output.status.success(),
-        "command failed with stderr: {}",
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    (
-        String::from_utf8_lossy(&output.stdout).to_string(),
-        String::from_utf8_lossy(&output.stderr).to_string(),
-    )
-}
-
-fn command_with_home(home: &Path) -> Command {
-    let mut cmd = Command::cargo_bin("lsp").unwrap();
-    cmd.env("HOME", home);
-    cmd
-}
-
-fn has_ansi(text: &str) -> bool {
-    text.contains("\u{1b}[")
-}
+use common::{
+    command_with_home, has_ansi, run_and_capture, run_and_capture_raw,
+};
 
 #[test]
 fn test_version_flag() {
