+++ tests/integration_unix.rs
@@ -5,9 +5,14 @@ use lsplus::utils::icons::Icon;
 use nix::unistd::Uid;
 use std::fs;
 use std::os::unix::fs::PermissionsExt;
-use std::path::Path;
 use strip_ansi_escapes::strip_str;
 use tempfile::tempdir;
+
+mod common;
+
+use common::{
+    command_with_home, has_ansi, run_and_capture, run_and_capture_raw,
+};
 
 struct PermissionGuard {
     path: std::path::PathBuf,
@@ -18,43 +23,6 @@ impl Drop for PermissionGuard {
         let _ =
             fs::set_permissions(&self.path, fs::Permissions::from_mode(0o700));
     }
-}
-
-fn run_and_capture(cmd: &mut Command) -> (String, String) {
-    let output = cmd.output().unwrap();
-    assert!(
-        output.status.success(),
-        "command failed with stderr: {}",
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    let stdout = strip_str(String::from_utf8_lossy(&output.stdout)).to_owned();
-    let stderr = strip_str(String::from_utf8_lossy(&output.stderr)).to_owned();
-    (stdout, stderr)
-}
-
-fn run_and_capture_raw(cmd: &mut Command) -> (String, String) {
-    let output = cmd.output().unwrap();
-    assert!(
-        output.status.success(),
-        "command failed with stderr: {}",
-        String::from_utf8_lossy(&output.stderr)
-    );
-
-    (
-        String::from_utf8_lossy(&output.stdout).to_string(),
-        String::from_utf8_lossy(&output.stderr).to_string(),
-    )
-}
-
-fn command_with_home(home: &Path) -> Command {
-    let mut cmd = Command::cargo_bin("lsp").unwrap();
-    cmd.env("HOME", home);
-    cmd
-}
-
-fn has_ansi(text: &str) -> bool {
-    text.contains("\u{1b}[")
 }
 
 fn create_indicator_fixture() -> tempfile::TempDir {
